Support arbitrary size integers
This adds support for arbitrarily large integers and arbitrary precision
floating point numbers.

serde_json has some support for this feature which makes it doable by
turning on a feature and doing a bit of trickery in the visitor. The
serializer path is more straightforward as you can use the repr of an
integer to get a string to pass directly to serde to use.

I don't think this will have a major impact on the performance in the
normal case as serde handles 64-bit values by going through a faster
path and not doing the string inside a map dance. That said, this will
have some impact on performance but I haven't measured it yet. Also,
this will serialize and deserialize some things that were previously
invalid.